+++ packages/preview/CHANGELOG.md
@@ -1,8 +1,15 @@
 # @shermant/i18n-translator-preview
 
+## 1.0.1
+
+### Minor Changes
+
+- Add README docs
+- Update example file synchronously
+
 ## 1.0.0
 
 ### Major Changes
 
 - Change vite build baseUrl
-- e846782: Add function previewer
+- Add function previewer